Refactor code, arrange files
