+++ Web development/Website Project/script.js
